backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
pytesseract==0.3.10
pdf2image==1.17.0
pandas==2.2.2
openpyxl==3.1.2
python-dateutil==2.9.0
httpx==0.27.0
